Add files via upload
Examples of two generated patients with Synthea in FHIR format, along with csvs considering observations and conditions organized by date (for potential use in a model). Csvs are forward filled, with binary values for conditions and numerical (unprocessed) values for observations. NAN values are filled with zeroes, however for continuous features this may be changed dynamically for modeling purposes.